Starting three tasks
Three tasks to use in comparison.
Summing to sum natural numbers up to n. Will be tested on several values up to 1e10. Currrently done only on Java.

Unoptimal recursive Fibonacci. Original idea was to test whether Haskell cache results or not. Done on languages.

Program to find derivative of given equation. Used to show ease of function pattern matching in Haskell. Done on Haskell and half-way for Java.